requirements.txt
# Core dependencies
numpy>=1.24.0
networkx>=3.1
pandas>=2.0.0

# LLM and Embeddings
openai>=1.0.0
tiktoken>=0.5.0

# LightRAG
lightrag-hku>=1.2.0

# Vector storage
faiss-cpu>=1.7.4

# Evaluation
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
jupyter>=1.0.0
